refactor(content-parser): move extractFoodItemsFromContent to legacy/, delete dead parseMealFromMarkdown
extractFoodItemsFromContent moves to src/utils/legacy/parser.ts, used only
by the upcoming v2->v3 migration to read pre-3.0 food log notes.
parseMealFromMarkdown is deleted outright rather than moved — its only two
callers (syncMealNoteToJSON, the old handleFileRename) were both removed in
Phase 2c, and meal items were always JSON-backed so migration never needed
it.

Also adds PluginSettings.dataVersion, deliberately absent from
DEFAULT_SETTINGS so the migration can tell a pre-3.0 vault (field missing
from saved data.json) apart from one that's already migrated.
